Moves most enums to the models package (#188)
