+++ apiTester/src/commonMain/kotlin/com/revenuecat/purchases/kmp/apitester/ProductTypeAPI.kt
@@ -1,6 +1,6 @@
 package com.revenuecat.purchases.kmp.apitester
 
-import com.revenuecat.purchases.kmp.ProductType
+import com.revenuecat.purchases.kmp.models.ProductType
 
 @Suppress("unused")
 private class ProductTypeAPI {
