+++ apiTester/src/commonMain/kotlin/com/revenuecat/purchases/kmp/apitester/PurchasesAreCompletedByAPI.kt
@@ -1,7 +1,7 @@
 package com.revenuecat.purchases.kmp.apitester
 
 import com.revenuecat.purchases.kmp.PurchasesAreCompletedBy
-import com.revenuecat.purchases.kmp.StoreKitVersion
+import com.revenuecat.purchases.kmp.models.StoreKitVersion
 
 @Suppress("unused", "UNUSED_VARIABLE")
 private class PurchasesAreCompletedByAPI {
