+++ apiTester/src/commonMain/kotlin/com/revenuecat/purchases/kmp/apitester/StoreKitVersionAPI.kt
@@ -1,6 +1,6 @@
 package com.revenuecat.purchases.kmp.apitester
 
-import com.revenuecat.purchases.kmp.StoreKitVersion
+import com.revenuecat.purchases.kmp.models.StoreKitVersion
 
 @Suppress("unused")
 private class StoreKitVersionAPI {
