+++ core/src/commonMain/kotlin/com/revenuecat/purchases/kmp/Purchases.kt
@@ -2,6 +2,7 @@ package com.revenuecat.purchases.kmp
 
 import com.revenuecat.purchases.kmp.PurchasesConfiguration.Builder
 import com.revenuecat.purchases.kmp.models.BillingFeature
+import com.revenuecat.purchases.kmp.models.CacheFetchPolicy
 import com.revenuecat.purchases.kmp.models.PromotionalOffer
 import com.revenuecat.purchases.kmp.models.StoreMessageType
 import com.revenuecat.purchases.kmp.models.StoreProduct
