+++ either/api/either.klib.api
@@ -12,7 +12,7 @@ final class com.revenuecat.purchases.kmp.either/FailedPurchase { // com.revenuec
     final val userCancelled // com.revenuecat.purchases.kmp.either/FailedPurchase.userCancelled|{}userCancelled[0]
         final fun <get-userCancelled>(): kotlin/Boolean // com.revenuecat.purchases.kmp.either/FailedPurchase.userCancelled.<get-userCancelled>|<get-userCancelled>(){}[0]
 }
-final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.either/awaitCustomerInfoEither(com.revenuecat.purchases.kmp/CacheFetchPolicy =...): arrow.core/Either<com.revenuecat.purchases.kmp/PurchasesError, com.revenuecat.purchases.kmp/CustomerInfo> // com.revenuecat.purchases.kmp.either/awaitCustomerInfoEither|awaitCustomerInfoEither@com.revenuecat.purchases.kmp.Purchases(com.revenuecat.purchases.kmp.CacheFetchPolicy){}[0]
+final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.either/awaitCustomerInfoEither(com.revenuecat.purchases.kmp.models/CacheFetchPolicy =...): arrow.core/Either<com.revenuecat.purchases.kmp/PurchasesError, com.revenuecat.purchases.kmp/CustomerInfo> // com.revenuecat.purchases.kmp.either/awaitCustomerInfoEither|awaitCustomerInfoEither@com.revenuecat.purchases.kmp.Purchases(com.revenuecat.purchases.kmp.models.CacheFetchPolicy){}[0]
 final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.either/awaitGetProductsEither(kotlin.collections/List<kotlin/String>): arrow.core/Either<com.revenuecat.purchases.kmp/PurchasesError, kotlin.collections/List<com.revenuecat.purchases.kmp.models/StoreProduct>> // com.revenuecat.purchases.kmp.either/awaitGetProductsEither|awaitGetProductsEither@com.revenuecat.purchases.kmp.Purchases(kotlin.collections.List<kotlin.String>){}[0]
 final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.either/awaitLogInEither(kotlin/String): arrow.core/Either<com.revenuecat.purchases.kmp/PurchasesError, com.revenuecat.purchases.kmp.ktx/SuccessfulLogin> // com.revenuecat.purchases.kmp.either/awaitLogInEither|awaitLogInEither@com.revenuecat.purchases.kmp.Purchases(kotlin.String){}[0]
 final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.either/awaitLogOutEither(): arrow.core/Either<com.revenuecat.purchases.kmp/PurchasesError, com.revenuecat.purchases.kmp/CustomerInfo> // com.revenuecat.purchases.kmp.either/awaitLogOutEither|awaitLogOutEither@com.revenuecat.purchases.kmp.Purchases(){}[0]
