+++ mappings/src/androidMain/kotlin/com/revenuecat/purchases/kmp/mappings/ProductType.android.kt
@@ -1,6 +1,6 @@
 package com.revenuecat.purchases.kmp.mappings
 
-import com.revenuecat.purchases.kmp.ProductType
+import com.revenuecat.purchases.kmp.models.ProductType
 import com.revenuecat.purchases.ProductType as AndroidProductType
 
 internal fun AndroidProductType.toProductType(): ProductType {
