+++ mappings/src/iosMain/kotlin/com/revenuecat/purchases/kmp/mappings/CacheFetchPolicy.ios.kt
@@ -4,7 +4,7 @@ import cocoapods.PurchasesHybridCommon.RCCacheFetchPolicyCachedOrFetched
 import cocoapods.PurchasesHybridCommon.RCCacheFetchPolicyFetchCurrent
 import cocoapods.PurchasesHybridCommon.RCCacheFetchPolicyFromCacheOnly
 import cocoapods.PurchasesHybridCommon.RCCacheFetchPolicyNotStaleCachedOrFetched
-import com.revenuecat.purchases.kmp.CacheFetchPolicy
+import com.revenuecat.purchases.kmp.models.CacheFetchPolicy
 import cocoapods.PurchasesHybridCommon.RCCacheFetchPolicy as IosCacheFetchPolicy
 
 public fun CacheFetchPolicy.toIosCacheFetchPolicy(): IosCacheFetchPolicy =
