+++ mappings/src/iosMain/kotlin/com/revenuecat/purchases/kmp/mappings/VerificationResult.ios.kt
@@ -1,6 +1,6 @@
 package com.revenuecat.purchases.kmp.mappings
 
-import com.revenuecat.purchases.kmp.VerificationResult
+import com.revenuecat.purchases.kmp.models.VerificationResult
 import cocoapods.PurchasesHybridCommon.RCVerificationResult as IosVerificationResult
 import cocoapods.PurchasesHybridCommon.RCVerificationResultFailed as IosVerificationResultFailed
 import cocoapods.PurchasesHybridCommon.RCVerificationResultNotRequested as IosVerificationResultNotRequested
