+++ mappings/src/iosTest/kotlin/com/revenuecat/purchases/kmp/mappings/StoreKitVersionTests.ios.kt
@@ -1,6 +1,6 @@
 package com.revenuecat.purchases.kmp.mappings
 
-import com.revenuecat.purchases.kmp.StoreKitVersion
+import com.revenuecat.purchases.kmp.models.StoreKitVersion
 import kotlin.test.Test
 import kotlin.test.assertEquals
