+++ models/src/commonMain/kotlin/com/revenuecat/purchases/kmp/models/CacheFetchPolicy.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.kmp
+package com.revenuecat.purchases.kmp.models
 
 /**
  * Specifies behavior for a caching API.
