+++ models/src/commonMain/kotlin/com/revenuecat/purchases/kmp/models/EntitlementVerificationMode.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.kmp
+package com.revenuecat.purchases.kmp.models
 
 /**
  * Verification strictness levels for [EntitlementInfo].
