+++ models/src/commonMain/kotlin/com/revenuecat/purchases/kmp/models/ProductType.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.kmp
+package com.revenuecat.purchases.kmp.models
 
 /**
  * Possible product types.
