+++ models/src/commonMain/kotlin/com/revenuecat/purchases/kmp/models/PurchasingData.kt
@@ -1,7 +1,5 @@
 package com.revenuecat.purchases.kmp.models
 
-import com.revenuecat.purchases.kmp.ProductType
-
 /**
  * Data connected to a purchase.
  */
