+++ models/src/commonMain/kotlin/com/revenuecat/purchases/kmp/models/StoreKitVersion.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.kmp
+package com.revenuecat.purchases.kmp.models
 
 /**
  * Defines which version of StoreKit may be used.
