+++ result/api/result.klib.api
@@ -6,7 +6,7 @@
 // - Show declarations: true
 
 // Library unique name: <com.revenuecat.purchases:result>
-final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.result/awaitCustomerInfoResult(com.revenuecat.purchases.kmp/CacheFetchPolicy =...): kotlin/Result<com.revenuecat.purchases.kmp/CustomerInfo> // com.revenuecat.purchases.kmp.result/awaitCustomerInfoResult|awaitCustomerInfoResult@com.revenuecat.purchases.kmp.Purchases(com.revenuecat.purchases.kmp.CacheFetchPolicy){}[0]
+final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.result/awaitCustomerInfoResult(com.revenuecat.purchases.kmp.models/CacheFetchPolicy =...): kotlin/Result<com.revenuecat.purchases.kmp/CustomerInfo> // com.revenuecat.purchases.kmp.result/awaitCustomerInfoResult|awaitCustomerInfoResult@com.revenuecat.purchases.kmp.Purchases(com.revenuecat.purchases.kmp.models.CacheFetchPolicy){}[0]
 final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.result/awaitGetProductsResult(kotlin.collections/List<kotlin/String>): kotlin/Result<kotlin.collections/List<com.revenuecat.purchases.kmp.models/StoreProduct>> // com.revenuecat.purchases.kmp.result/awaitGetProductsResult|awaitGetProductsResult@com.revenuecat.purchases.kmp.Purchases(kotlin.collections.List<kotlin.String>){}[0]
 final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.result/awaitLogInResult(kotlin/String): kotlin/Result<com.revenuecat.purchases.kmp.ktx/SuccessfulLogin> // com.revenuecat.purchases.kmp.result/awaitLogInResult|awaitLogInResult@com.revenuecat.purchases.kmp.Purchases(kotlin.String){}[0]
 final suspend fun (com.revenuecat.purchases.kmp/Purchases).com.revenuecat.purchases.kmp.result/awaitLogOutResult(): kotlin/Result<com.revenuecat.purchases.kmp/CustomerInfo> // com.revenuecat.purchases.kmp.result/awaitLogOutResult|awaitLogOutResult@com.revenuecat.purchases.kmp.Purchases(){}[0]
